Refactor Homebrew setup
This is a complete rewrite which provides easier maintainability. The
`install.sh` file reads from the text files in the /brew folder and
installs the listed packages. Managing packages is as easy as
adding/removing packages form the files, or simply piping the output of
`brew list`/`brew cask list` to the appropriate file.